+++ R/feasts.R
@@ -6,6 +6,6 @@ globalVariables(".")
 #' @import rlang
 #' @import tsibble
 #' @importFrom dplyr mutate transmute summarise group_by ungroup filter distinct lead
-#' @importFrom tidyr nest unnest gather
+#' @importFrom tidyr gather
 NULL