+++ R/utils.R
@@ -27,6 +27,31 @@ lag <- function(x, n){
   out
 }
 
+nest_keys <- function(.data, nm = "data"){
+  out <- unclass(key_data(.data))
+  key <- key_vars(.data)
+  row_indices <- out[[length(out)]]
+  out[[length(out)]] <- NULL
+  col_nest <- -match(key, colnames(.data))
+  if(is_empty(col_nest)){
+    col_nest <- NULL
+  }
+  idx <- index_var(.data)
+  idx2 <- index2_var(.data)
+  ordered <- is_ordered(.data)
+  regular <- is_regular(.data)
+  out[[nm]] <- map(row_indices, function(x, i, j){
+    out <- if(is.null(j)) x[i,] else x[i,j]
+    build_tsibble_meta(
+      out,
+      key_data = as_tibble(list(.rows = list(seq_along(i)))),
+      index = idx, index2 = idx2, ordered = ordered,
+      interval = if(length(i) > 1 && regular) interval_pull(out[[idx]]) else interval(.data)
+    )
+  }, x = as_tibble(.data), j = col_nest)
+  as_tibble(out)
+}
+
 unnest_tbl <- function(.data, tbl_col, .sep = NULL){
   row_indices <- rep.int(seq_len(NROW(.data)), map_int(.data[[tbl_col[[1]]]], NROW))